Change all lpi and lpk Lazarus files to always use stabs debug format. This fixes Linux x86-64 + FPC 2.6.0 + Lazarus problems: fpc 2.6.0 (only on Linux x86-64, it seems) makes Access Violations if you mix debug format (some parts compiled with -g, some with -gs). Although this is pure FPC bug, it was triggered only when compiling from Lazarus/lazbuild, because some of our lpk/lpi files had
    <Linking>
      <Debugging>
	<GenerateDebugInfo Value="True"/>
      </Debugging>
    <Linking>

(resulting in -gs in Lazarus) and some did not (resulting only in -g).

This was causing Access Violations when trying to compile castle_components package, or fps_game project, or model_3d_viewer, and probably some others too, with lazbuild.

Things get even more complicated as defaults whether to generate debug info, and whether default is stabs (-gs) or platform default (-g) changed in Lazarus, depending on

  <CompilerOptions>
    <Version Value="11"/>

For version < 11 (most of our files for now) we need <GenerateDebugInfo Value="True"/>, and it will result in stabs. For version >= 11, we need 

    <Linking>
      <Debugging>
        <DebugInfoType Value="dsStabs"/>
      </Debugging>
    </Linking>


git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/glplotter@12361 afe8dcc5-9094-451b-8773-5f495042a737