Cleanup (deprecation, will be removed in next release) of some outdated CastleUtils dialog boxes routines.
This stuff is ugly, and doesn't really belong to CastleUtils unit. It's there only for "emergency" methods to display error messages in case of Windows GUI programs, and should not be used in normal circumstances. So, mark most of them as deprecated, and suggest to rather use LCL dialogs (inside LCL) or CastleWindow dialogs and controls.

git-svn-id: https://svn.code.sf.net/p/castle-engine/code/trunk/glplotter@12945 afe8dcc5-9094-451b-8773-5f495042a737